refactor: Syntaxis, code-style, and more :p
